+++ Assets/Animations/Fly Fly.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fbf8b78073933940b01c4ef8c95a76f
+timeCreated: 1482612559
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
