+++ Assets/Animations/Fly.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f10d11e7ccece514fafafaa5bd76df92
+timeCreated: 1482612559
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
