+++ Assets/Animations/Snake Crawl.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f880df1f39818b438fa8a73d05b7c23
+timeCreated: 1482613264
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
