+++ Assets/Animations/snake.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22c2563a1a798f04db7eebe106dfe167
+timeCreated: 1482613265
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
