+++ Assets/Prefabs/Enemies/Enemy Fly.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 336b164ac862b414d9f10f0aea8cf0ec
+timeCreated: 1482612895
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
