+++ Assets/Prefabs/Enemies/Enemy Snake.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3229e1060ee88346b0166f2682df4e5
+timeCreated: 1482613554
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
