+++ Assets/Prefabs/Level Pieces/Moving Platform Purple.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac41d2a3fce262f469b6c741199f4065
+timeCreated: 1482611865
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
