+++ Assets/Scripts/MoveController02.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 2d94e34e3921988419cf73ef5153143f
+timeCreated: 1479556716
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
